https://github.com/jackdewinter/pymarkdown/issues/589  - #closes 589
