+++ changelog.md
@@ -8,7 +8,8 @@
 
 ### Changed
 
-- None
+- [Issue 582](https://github.com/jackdewinter/pymarkdown/issues/582)
+  - broke up leaf block processor into more distinct pieces
 
 ### Fixed
