+++ pymarkdown/container_block_leaf_processor.py
@@ -36,7 +36,6 @@ from pymarkdown.tab_helper import TabHelper
 
 POGGER = ParserLogger(logging.getLogger(__name__))
 
-# pylint: disable=too-many-lines
 # pylint: disable=too-few-public-methods
